Extract BoardPoller from IssueTracker (CROW-1094)
Move the board poll into its own collaborator: the consolidated GitHub GraphQL
query, stale-PR follow-up fetch, GitLab/Jira assigned-issue fetches, PR-record
dedup (mergePRRecords/dedupedByURL/etc.), and the crow:auto dispatch. It owns
the auto-create in-flight / log-rate-limit bookkeeping and reaches the shared
warnings+rate-limit sink (handleGitHubBackendError, report/clearScope/SAMLWarning
— widened to internal, kept with the warnings concern), the auto-create
callback/toggle, appState, providerManager, and the shared JSONStore through an
unowned back-reference. Jira config resolution (jiraConfig/jiraStatusMap/
matchJiraWorkspace) stays with the ticket-transition API it serves.

refresh() drives every BoardPoller method in the current order; an
extension IssueTracker re-exposes the IssueTracker.<symbol> spelling (the shared
PR-record helpers, GitLab parse statics, mergeListing, canRunAutoCreate,
Consolidated/StalePR response types) used by the other collaborators and tests.

swift test --package-path Packages/CrowEngine: 789 tests pass.


Crow-Session: 18A4CD55-8B33-48E4-B5EB-E710314DF5E5
